feat: add global clickable live users badge
- Make LiveUsersBadge clickable with link to /live page
- Add hover effects and "Click to view details" tooltip text
- Move badge to global _app.tsx layout for display on all pages
- Remove duplicate badge from home page index route
- Add /live link to home page navigation grid
