+++ routes/_app.tsx
@@ -1,6 +1,34 @@
-import { type PageProps } from "$fresh/server.ts";
+import { FreshContext, Handlers, type PageProps } from "$fresh/server.ts";
+import { kv } from "@/libs/kv.ts";
+import { LiveSessionManager, LiveStats } from "@/libs/live-sessions.ts";
+import LiveUsersBadge from "@/islands/LiveUsersBadge.tsx";
 
-export default function App({ Component }: PageProps) {
+interface AppData {
+  liveStats: LiveStats;
+}
+
+export const handler: Handlers<AppData> = {
+  async GET(_req: Request, ctx: FreshContext) {
+    const sessionManager = new LiveSessionManager(kv);
+
+    try {
+      const liveStats = await sessionManager.getLiveStats();
+      return ctx.render({ liveStats });
+    } catch (error) {
+      console.error("Failed to get live stats in app:", error);
+      return ctx.render({
+        liveStats: {
+          total: 0,
+          byType: { human: 0, bot: 0, llm: 0 },
+          byLocation: {},
+          trend: [],
+        },
+      });
+    }
+  },
+};
+
+export default function App({ Component, data }: PageProps<AppData>) {
   return (
     <html>
       <head>
@@ -13,6 +41,15 @@ export default function App({ Component }: PageProps) {
       </head>
       <body>
         <Component />
+        <LiveUsersBadge
+          initialStats={data?.liveStats ||
+            {
+              total: 0,
+              byType: { human: 0, bot: 0, llm: 0 },
+              byLocation: {},
+              trend: [],
+            }}
+        />
       </body>
     </html>
   );
